更改 Resolution 类型为 float 并添加宏定义
在 Shader.hlsl 文件中：
* 添加了三个宏定义：NUM_AA、NUM_MIN 和 NUM_MAX。
* 将 Constants 结构体中 Resolution 的类型从 uint2 改为 float2。
* 简化了 RayMarching 函数中 uv 的计算。

在 ComputeShaderTest.cs 文件中：
* 将 Constants 结构体中 Resolution 的类型从 Vector2D<uint> 改为 Vector2D<float>。

这些更改主要是为了提高计算精度，改进渲染效果，并提高代码的可读性和简洁性。

+++ src/Examples/ComputeShader/ComputeShaderTest.cs
@@ -16,7 +16,7 @@ internal unsafe class ComputeShaderTest(Backend backend) : VisualTest("Compute S
     private struct Constants
     {
         [FieldOffset(0)]
-        public Vector2D<uint> Resolution;
+        public Vector2D<float> Resolution;
 
         [FieldOffset(8)]
         public float DeltaTime;
